# libibverbs: Speed up MR registration when RDMAV_HUGEPAGES_SAFE is ena…