…bled

When RDMAV_HUGEPAGES_SAFE is enabled, every MR registration has an extra
overhead as a result of searching for the appropriate page size backing
up the buffer.

Since looking for the page size by parsing /proc/pid/smaps takes a
considerable time, this patch speeds it up by calling madvise with the
various page sizes until one of the calls succeeds (with the correct
page size).

Tested by measuring average time to complete ibv_reg_mr:
Before this patch:
RDMAV_HUGEPAGES_SAFE unset: 93us
RDMAV_HUGEPAGES_SAFE set  : 5414us (58x)

After this patch:
RDMAV_HUGEPAGES_SAFE unset: 93us
RDMAV_HUGEPAGES_SAFE set  : 103us (1.1x)

The possible page sizes are queried once on ibv_fork_init from the
/sys/kernel/mm/hugepages directory.
The get_page_size function iterates the different page sizes and calls
madvise for each, first with DOFORK and then with DONTFORK.

This is done for two reasons:
First, it's correct to end up with the pages set to DONTFORK as the
rbtree will later handle the reference counting and set the advice for
these pages appropriately.

Second, the kernel caches the vma flags and doesn't check madvise's
address and length parameters if the flags did not change. This means
that if the same page is being madvise'd twice the second call will
always succeed regardless of the page size. In order to solve that the
initial DOFORK was added.

Suggested-by: Wei Zhang <wzam@amazon.com>
Signed-off-by: Gal Pressman <galpress@amazon.com>