@galpress 2020-06-01
This patch feels less intuitive than the existing code, but I think it's worth considering it as it shows a significant improvement over the current state.

The DOFORK -> DONTFORK behavior might be a bit problematic, but it can be changed to MADV_SEQUENTIAL -> MADV_NORMAL for example.